+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## wdio-intercept-service changelog
 
+### [ [>](https://github.com/chmanie/wdio-intercept-service/tree/vnext) ] next / xx.xx.xxxx
+* Add full typescript support for the plugin
+
 ### [ [>](https://github.com/chmanie/wdio-intercept-service/tree/v3.1.1) ] 3.1.1 / 11.01.2020
 * Fix incorrect check for assertExpectedRequestsOnly (thanks @chrisdraycott)